requirements.txt
streamlit
pandas
openreview-py
sentence-transformers
scikit-learn
google-generativeai
openai
numpy
tqdm
googlesearch-python
pypdf
requests